+++ esa-20240910/core/Models/GetRoutineResponseBody.cs
@@ -83,16 +83,6 @@ namespace AlibabaCloud.SDK.ESA20240910.Models
             [Validation(Required=false)]
             public string Env { get; set; }
 
-            /// <summary>
-            /// <para>The specification.</para>
-            /// 
-            /// <b>Example:</b>
-            /// <para>5ms</para>
-            /// </summary>
-            [NameInMap("SpecName")]
-            [Validation(Required=false)]
-            public string SpecName { get; set; }
-
         }
 
         /// <summary>