+++ esa-20240910/core/Models/ListRoutineRelatedRecordsRequest.cs
@@ -1,0 +1,48 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.ESA20240910.Models
+{
+    public class ListRoutineRelatedRecordsRequest : TeaModel {
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>test</para>
+        /// </summary>
+        [NameInMap("Name")]
+        [Validation(Required=false)]
+        public string Name { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1</para>
+        /// </summary>
+        [NameInMap("PageNumber")]
+        [Validation(Required=false)]
+        public long? PageNumber { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>20</para>
+        /// </summary>
+        [NameInMap("PageSize")]
+        [Validation(Required=false)]
+        public long? PageSize { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>hello</para>
+        /// </summary>
+        [NameInMap("SearchKeyWord")]
+        [Validation(Required=false)]
+        public string SearchKeyWord { get; set; }
+
+    }
+
+}