+++ esa-20240910/core/Models/ListRoutineRelatedRecordsResponseBody.cs
@@ -1,0 +1,84 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.ESA20240910.Models
+{
+    public class ListRoutineRelatedRecordsResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1</para>
+        /// </summary>
+        [NameInMap("PageNumber")]
+        [Validation(Required=false)]
+        public long? PageNumber { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>20</para>
+        /// </summary>
+        [NameInMap("PageSize")]
+        [Validation(Required=false)]
+        public long? PageSize { get; set; }
+
+        [NameInMap("RelatedRecords")]
+        [Validation(Required=false)]
+        public List<ListRoutineRelatedRecordsResponseBodyRelatedRecords> RelatedRecords { get; set; }
+        public class ListRoutineRelatedRecordsResponseBodyRelatedRecords : TeaModel {
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>509348423011904</para>
+            /// </summary>
+            [NameInMap("RecordId")]
+            [Validation(Required=false)]
+            public long? RecordId { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>test-record-1.example.com</para>
+            /// </summary>
+            [NameInMap("RecordName")]
+            [Validation(Required=false)]
+            public string RecordName { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>54362329990032</para>
+            /// </summary>
+            [NameInMap("SiteId")]
+            [Validation(Required=false)]
+            public long? SiteId { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>example.com</para>
+            /// </summary>
+            [NameInMap("SiteName")]
+            [Validation(Required=false)]
+            public string SiteName { get; set; }
+
+        }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>15C66C7B-671A-4297-9187-2C4477247A74</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>16</para>
+        /// </summary>
+        [NameInMap("TotalCount")]
+        [Validation(Required=false)]
+        public long? TotalCount { get; set; }
+
+    }
+
+}