+++ esa-20240910/core/Models/ListUserRoutinesRequest.cs
@@ -1,0 +1,30 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.ESA20240910.Models
+{
+    public class ListUserRoutinesRequest : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>ListUserRoutines</para>
+        /// </summary>
+        [NameInMap("PageNumber")]
+        [Validation(Required=false)]
+        public long? PageNumber { get; set; }
+
+        [NameInMap("PageSize")]
+        [Validation(Required=false)]
+        public long? PageSize { get; set; }
+
+        [NameInMap("SearchKeyWord")]
+        [Validation(Required=false)]
+        public string SearchKeyWord { get; set; }
+
+    }
+
+}